+++ Userland/Libraries/LibGfx/Font/ScaledFont.cpp
@@ -10,12 +10,12 @@
 
 namespace Gfx {
 
-int ScaledFont::width(StringView view) const { return unicode_view_width(Utf8View(view)); }
-int ScaledFont::width(Utf8View const& view) const { return unicode_view_width(view); }
-int ScaledFont::width(Utf32View const& view) const { return unicode_view_width(view); }
+float ScaledFont::width(StringView view) const { return unicode_view_width(Utf8View(view)); }
+float ScaledFont::width(Utf8View const& view) const { return unicode_view_width(view); }
+float ScaledFont::width(Utf32View const& view) const { return unicode_view_width(view); }
 
 template<typename T>
-ALWAYS_INLINE int ScaledFont::unicode_view_width(T const& view) const
+ALWAYS_INLINE float ScaledFont::unicode_view_width(T const& view) const
 {
     if (view.is_empty())
         return 0;
